requirements.txt
boto3==1.38.7
botocore==1.38.7
certifi==2025.4.26
charset-normalizer==3.4.2
cramjam==2.10.0
docker==7.1.0
exceptiongroup==1.2.2
fastparquet==2024.11.0
fsspec==2025.3.2
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
testcontainers==4.10.0
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wrapt==1.17.2
